+++ settings.js
@@ -149,7 +149,7 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
     category: "Chat",
     configName: "messageOnTheLochEmperorCatch",
     title: "Send a party chat message on THE LOCH EMPEROR catch",
-    description: `Sends a party chat message when a rare sea creature has caught by you.\n${RED}Unavailable until the module is migrated to work on MC 1.21!`,
+    description: `Sends a party chat message when a rare sea creature has caught by you.`,
     subcategory: "Rare Catches",
     value: true
 })
@@ -329,7 +329,7 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
 .addSwitch({
     category: "Chat",
     configName: "messageOnBabyYetiPetDrop",
-    title: "Send a party chat message on BABY YETI PET drop",
+    title: "Send a party chat message on LEGENDARY BABY YETI PET drop",
     description: "Sends a party chat message when a rare item has dropped by you.",
     subcategory: "Rare Drops",
     value: true
@@ -337,7 +337,7 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
 .addSwitch({
     category: "Chat",
     configName: "messageOnFlyingFishPetDrop",
-    title: "Send a party chat message on FLYING FISH PET drop",
+    title: "Send a party chat message on LEGENDARY FLYING FISH PET drop",
     description: "Sends a party chat message when a rare item has dropped by you.",
     subcategory: "Rare Drops",
     value: true
@@ -410,6 +410,14 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
     category: "Chat",
     configName: "messageOnSquidPetDrop",
     title: "Send a party chat message on SQUID PET drop",
+    description: "Sends a party chat message when a rare item has dropped by you.",
+    subcategory: "Rare Drops",
+    value: true
+})
+.addSwitch({
+    category: "Chat",
+    configName: "messageOnPhoenixPetDrop",
+    title: "Send a party chat message on PHOENIX PET drop",
     description: "Sends a party chat message when a rare item has dropped by you.",
     subcategory: "Rare Drops",
     value: true
@@ -663,6 +671,15 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
 })
 .addSlider({
     category: "Alerts",
+    configName: "seaCreaturesCountThreshold_Galatea",
+    title: "Sea creatures count threshold - GALATEA",
+    description: "Count of sea creatures nearby required to see the alert when you are in the Galatea. Ignored if the sea creatures count alert is disabled.",
+    options: [5, 60],
+    value: 30,
+    subcategory: "Sea creatures count"
+})
+.addSlider({
+    category: "Alerts",
     configName: "seaCreaturesCountThreshold_Default",
     title: "Sea creatures count threshold - Other",
     description: "Count of sea creatures nearby required to see the alert when you are in other locations. Ignored if the sea creatures count alert is disabled.",
@@ -729,7 +746,7 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
     category: "Alerts",
     configName: "alertOnTheLochEmperorCatch",
     title: "Alert on THE LOCH EMPEROR catch",
-    description: `Shows a title and plays a sound when a rare sea creature has caught by you or your party members.\n${RED}Unavailable until the module is migrated to work on MC 1.21!`,
+    description: `Shows a title and plays a sound when a rare sea creature has caught by you or your party members.`,
     subcategory: "Rare Catches",
     value: true
 })
@@ -857,7 +874,7 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
 .addSwitch({
     category: "Alerts",
     configName: "alertOnBabyYetiPetDrop",
-    title: "Alert on BABY YETI PET drop",
+    title: "Alert on LEGENDARY BABY YETI PET drop",
     description: "Shows a title and plays a sound when a rare item has dropped by you or your party members.",
     subcategory: "Rare Drops",
     value: true
@@ -865,7 +882,7 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
 .addSwitch({
     category: "Alerts",
     configName: "alertOnFlyingFishPetDrop",
-    title: "Alert on FLYING FISH PET drop",
+    title: "Alert on LEGENDARY FLYING FISH PET drop",
     description: "Shows a title and plays a sound when a rare item has dropped by you or your party members.",
     subcategory: "Rare Drops",
     value: true
@@ -938,6 +955,14 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
     category: "Alerts",
     configName: "alertOnSquidPetDrop",
     title: "Alert on SQUID PET drop",
+    description: "Shows a title and plays a sound when a rare item has dropped by you or your party members.",
+    subcategory: "Rare Drops",
+    value: true
+})
+.addSwitch({
+    category: "Alerts",
+    configName: "alertOnPhoenixPetDrop",
+    title: "Alert on PHOENIX PET drop",
     description: "Shows a title and plays a sound when a rare item has dropped by you or your party members.",
     subcategory: "Rare Drops",
     value: true
@@ -1272,7 +1297,7 @@ const config = new DefaultConfig("FeeshNotifier", "config/settings.json")
     category: "Overlays",
     configName: "jerryWorkshopTrackerOverlay",
     title: "Jerry Workshop tracker",
-    description: `Shows an overlay with Yeti / Reindrake catch statistics and Baby Yeti pet drops statistics while in the Jerry Workshop.\nDo ${AQUA}/feeshResetJerryWorkshop${GRAY} to reset.`,
+    description: `Shows an overlay with Yeti / Reindrake catch statistics while in the Jerry Workshop.\nDo ${AQUA}/feeshResetJerryWorkshop${GRAY} to reset.`,
     subcategory: "Jerry Workshop tracker",
     value: true
 })
@@ -1724,7 +1749,7 @@ ${GRAY}Do ${AQUA}/feeshResetProfitTracker${GRAY} to reset.`,
     category: "Items and storages",
     configName: "showRarityUpgrade",
     title: "Rarity upgrade",
-    description: "Render rarity upgrade for recombobulated fishing items (autorecombobulator).",
+    description: "Render rarity upgrade flag (R) for auto-recombobulated fishing drops.",
     subcategory: "Item icon"
 })
 .addSwitch({